Tehtävät 4.15-4.21

+++ controllers/blogs.js
@@ -1,21 +1,47 @@
 const blogsRouter = require('express').Router()
+const jwt = require('jsonwebtoken')
 const Blog = require('../models/blog')
+const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog
+    .find({})
+    .populate('user', { 'username': 1, 'name': 1 })
+  response.json(blogs.map(Blog.format))
 })
 
 blogsRouter.post('/', async (request, response) => {
+  const body = request.body
+  const token = request.token
+
   try {
-    const blog = new Blog(request.body)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
+
     if(!blog.title || !blog.url) {
       return response.status(400).send({ error: 'title or url missing' })
     }
 
     if(!blog.likes) blog.likes = 0
     const savedBlog = await blog.save()
-    response.status(201).json(savedBlog)
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+    await user.save()
+
+    response.status(201).json(Blog.format(savedBlog))
   } catch (exception) {
     console.log(exception)
     response.status(500).json({ error: 'something went wrong...' })
@@ -23,7 +49,21 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = request.token
+
   try {
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+
+    if (!token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const blogToDelete = await Blog.findById(request.params.id)
+
+    if(blogToDelete.user.toString() !== decodedToken.id.toString()) {
+      return response.status(401).json({ error: 'can\'t delete blog created by another user' })
+    }
+
     await Blog.findByIdAndRemove(request.params.id)
     response.status(204).end()
   } catch (exception) {
